+++ jest.config.js
@@ -23,7 +23,7 @@ module.exports = {
         '^cnchar$': '<rootDir>/src/cnchar/main',
         '^cnchar-types$': '<rootDir>/src/cnchar-types',
         '^@cnchar-plugin/(.*?)$': '<rootDir>/src/cnchar/plugin/$1',
-        '^testUtils$': '<rootDir>/testUtils',
+        '^testUtils$': '<rootDir>/helper/testUtils',
     },
     testRegex: 'src/cnchar/main/__tests__/(.*).spec.ts?$',
     modulePathIgnorePatterns: [
