+++ tsconfig.json
@@ -26,11 +26,11 @@
             "cnchar": ["src/cnchar/main"],
             "cnchar-types": ["src/cnchar-types"],
             "@cnchar-plugin/*": ["src/cnchar/plugin/*"],
-            "testUtils":["testUtils"]
+            "testUtils":["helper/testUtils"]
         }
     },
     "include": [
-        "testUtils",
+        "helper/testUtils",
         "./src/**/*"   
     ],
     "exclude": [
